Support subflakes betterer
If you set the `upload-name` to whatever (i.e.
`NixOS/nixos-subflake-lib`), set the `repo` to the actual GitHub repo
(i.e. `NixOS/nixpkgs`), the `directory` (i.e. `./lib`), and finally the
`git_root` (i.e. `.`) to get a nxfr tarball that contained only the
Nixpkgs lib subflake.

This replaces the old `mirrored-for` with the `repo` input, but also
makes it more extensible / usable for e.g subflakes.