Read README from flake's store path
Before, we were reading the README from the current working directory
because of a couple things:

* We were reading `entry.file_name()` instead of `entry.path()`, which
would be something like `README.md`, instead of a full path to the file.
This would lead to reading the `README.md` (if it existed) from the
current directory rather than the directory we just searched through.

* We were using the `flake_root` (the root of the git repository) to
find the README. This would mean that we would read the `README.md` from
wherever this git root was, rather than from the flake itself. Now, we
read it from the flake's store path.