Fix compatibility with lazy trees
`nix flake metadata` does not return a `path` attribute when lazy
trees are enabled. We have to use `nix flake prefetch` to get a store
path. (Ideally Nix would have a command to dump a source tree to
stdout as a NAR or zip file so we don't have to copy it to the Nix
store...)